Coupler guard, front buttons, CamelCase enums, portability
- DeviceDisplay: guard processInputKo out on couplers (MASK_VERSION 0x09xx, e.g. IP-Router 0x091A) which have no GroupObjects
- ButtonManager: detect front-panel buttons via FRONT_CTRL_UP (native openknx.gpio / PCA9557) instead of USE_GPIO_MODULE
- WidgetsManager: rename WidgetManagerState/PowerSaveMode enum values to CamelCase (values unchanged)
- BootLogo: min((uint32_t)64, ...) fixes type-mismatched min()
- MatrixClassic: seed RNG with micros() (analogRead(0) is no valid ADC pin on ESP32-S3/RP2040)

+++ src/Widgets/BootLogo.cpp
@@ -42,7 +42,7 @@ void WidgetBootLogo::calculateOptimalStepHeight(uint32_t maxLoopTimeUs)
     if (_logoBitmap && _logoBitmap->width >= 100 && _logoBitmap->height >= 60)
     {
         // Fullscreen: Double the stripe size (less overhead)
-        calculatedStepHeight = min(64, calculatedStepHeight * 2);
+        calculatedStepHeight = min((uint32_t)64, calculatedStepHeight * 2);
     }
 
     // Clamp to reasonable values
